// types.d.ts
export type FormData = {
  name?: string;
  gender?: string;
  birth?: string;
  abroad_record?: boolean;
  bird_record?: boolean;
  fever_record?: boolean;
  education?: string;
  employment?: string;
  marriage?: string;
  living?: string;
  address?: string;
  alcohol_frequency?: string;
  smoking_frequency?: string;
  arrive_time?: string;
};